requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.0.0

